start add link + refactor
